Implementation of Firestore for Job Board
Job Board now retrieves data from firestore and displays
each job in a separate card on the Job Board. Functionality
achieved using a getJobs() function that queries a snapshot
of the data from the firestore database, and stores it into
the jobs array. The jobs array is then rendered onto the Job Board
page using a mapper function. The retrieval happens in real time,
which means that any change made on the firestore would automatically
update the information on the job board.