Changed C, U, D in CRUD for future development
In the future, we are planning to have more than 1 collection to
reference different types of data in the database. Changed
the reference for the create, update, and delete methods in UseStore
such that the collection type (e.g. "jobs") is parsed in as variable,
which would allow the same method to access multiple collections
on firestore. However, the read method is still singular as I have
yet to figure out how to mitigate the issue of the useEffect().