Refactor code and change naming convention
The URLs have been changed so that they are not underscore but hyphernated. The names of the URLs are also
tweaked such that more sign in pages for admins and students can be added without being confusing.
The login component was renamed to sign in. The landing page content was refactored into its own component.